src/ask_yogasutra/requirements.txt
networkx==3.2.1
rdflib==7.0.0
graphviz==0.20.1
matplotlib==3.8.2
pyvis==0.3.2
streamlit-agraph==0.0.45
pyarrow==14.0.2
streamlit==1.37.0
sentence-transformers==3.0.1
transformers==4.53.0
llama-index-core==0.10.57
llama-index-llms-groq==0.1.4
llama-index-embeddings-huggingface==0.2.2
llama-index-graph-stores-simple==0.2.0
langchain==0.2.10
psutil==5.9.8
groq==0.9.0
griffe==0.48.0
